Board save/load update and bug fixes
When deleting a column or card, the DB now updates to account for that upon saving. Before these objects wouldn't be removed from the db so when a user logged in again after saving, the columns or card they had deleted would still be loaded in. Fixed board loading to update mcolid (now a public variable).